Build libelf as a shared library.
The Linux kernel needs libelf.so to build, and as part of moving
to a hermetic toolchain to build the kernel, we need to build
libelf.so on the host.

This change introduces a non-default target libelf.so too, to
avoid having to introduce a new target name for the host. The
kernel build will only look for a file called 'libelf.so'.

Test: none
Bug: 80195544
Change-Id: I09918c0a6f94dc9d014250e73af204d267d0f204