Add build description for libdw
This adds build description for more elfutils libraries in order support
users that require libdw, such as pahole and libabigail.

The generic definition has been moved to elfutils_defaults to align
flags and visibility across the project.

While the remaining libraries could be considered standalone, they are
not really suitable to be built standalone in soong as they have
circular dependencies and are usually anyway built as one unit. Hence
only build libdw and include all libraries into it. This then works as
an extension for libelf.

To not introduce dependencies to m4, flex and bison, the libcpu platform
specific files have been generated and added as they would when using
the upstream tarball.

Change-Id: Ie6f0713812f10b5b12f4e0bdfca27d0369bb71a4
Signed-off-by: Matthias Maennich <maennich@google.com>